Handle macOS artifact name changes
I was playing around with `asdf` for installing the `purescript` `0.14.0` release candidates tonight, and I noticed the `asdf-spago` install wasn't working for spago `0.19.0`. I tried to patch the bash script in kind of a simple/hacky way to make it work for at least the last 5 or so versions of `spago`.

I tested `asdf install spago 0.16.0` and all versions up to the latest `0.19.0`. Before `0.16.0` the artifact names didn't seem to be consistent and/or the contents of the .tar.gz files did not contain the expected files, but I figured nobody would probably be trying to use those older versions at this point.